requirements.txt
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
websockets>=8.1
jsonschema>=3.2.0